requirements.txt

Flask==2.2.3
Werkzeug==2.2.3
flask-cors==3.0.10
google-cloud-dialogflow==2.13.0
gunicorn==20.1.0
